Refactor training results view into modular hooks
